+++ packages/openapi-validator/src/index.ts
@@ -1,0 +1,4 @@
+export { OpenApiVersions as supportedVersions } from '@/specifications'
+export type { ErrorObject, OpenApiVersion, ThrowOnErrorOption, ValidationOutcome } from '@/types'
+export { type ValidateOptions, validate } from '@/validate'
+export { validatePathParameters } from '@/validate-path-parameters'