+++ integrations/dotnet/azure-functions/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/azure-functions
 
+## 0.2.12
+
 ## 0.2.11
 
 ## 0.2.10