+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/blocks
 
+## 0.1.15
+
 ## 0.1.14
 
 ### Patch Changes