+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/core
 
+## 0.5.31
+
 ## 0.5.30
 
 ### Patch Changes